print-service/requirements.txt
requests
pillow
pywin32
dotenv
types-requests
types-pywin32